feature(estabBairro): Implementar endpoint de DELETE

+++ src/main/java/com/controle_de_gastos/notas_api/service/EstabelecimentoBairroJuncaoService.java
@@ -66,8 +66,11 @@ public class EstabelecimentoBairroJuncaoService {
         return toRespostaDTO(estabelecimentoBairroJuncaoRepository.save(ebJuncao));
     }
 
-    public void deletarPorId(Integer id){
-
+    public boolean deletarPorId(Integer id){
+        Optional<EstabelecimentoBairroJuncao> isDeletado = estabelecimentoBairroJuncaoRepository.findById(id);
+        if(isDeletado.isEmpty()) return false;
         estabelecimentoBairroJuncaoRepository.deleteById(id);
+        return true;
     }
+
 }
